requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
args
async-timeout
attrs
braceexpand
certifi
charset-normalizer
clin
clip
contourpy
cycler
datasets
diffusers
dill
einops
exceptiongroup
fairscale
filelock
fonttools
frozenlist
fsspec
ftfy
hpsv2
huggingface-hub
idna
image-reward==1.5
importlib_metadata
importlib_resources
iniconfig
ipykernel
ipython
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
multidict
multiprocess
networkx
numpy
pandas
pillow
pluggy
propcache
protobuf
pyarrow
pyparsing
pytest
pytest-split
PyYAML
regex
requests
safetensors
scipy
seaborn
sentencepiece
sympy
tokenizers
tomli
torch==2.4.0
torchvision==0.19.0
tqdm
transformers==4.38.2
triton==3.0.0
tzdata
urllib3
webdataset
xxhash
yarl
zipp
google-genai
typing_extensions